+++ neoforge/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 # [1.20.2-2.0.0.beta3] - UNRELEASED
 
 # Updated json files to reflect Neo Forges changes.
+# Identified issue with trident not being able to be thrown. Will be fixed in future.
 
 # [1.20.2-2.0.0.beta2] - 2023-11-13